Integrate PM2 into lisk.sh and lisk_snapshot.sh
In order to complete the migration to PM2 we must revise existing
shell scripts to accomodate the change. This set of changes
removes Forever and replaces it with PM2 in the relevate files.

The reason for the change in process manager is due to a bug in
Forever that sends SIGKILL regardless of the termination flag
specified by forever stop.