Correct regressions introduced by PM2
Prior to PM2 we would specify config.json as part of startup.
This file is now specified in PM2.json. There fore we need to
repoint the LISK_CONFIG var to the value in that file since
the old way no longer works to find the DB.

On startup Lisk generates a DAPP password if there is not one
present in the file. The sleep that allowed this to work was
removed in the PM2 migration causing a regression in the
installation of the software.

This fix corrects both of these issues by re adding the sleep
and changing the way the LISK_CONFIG var is set